Refactor menu editor forms into modal components
